#514 Separate whole detail page into 3 components
